+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Links in the changelog are part of [github.com/Jelmerro/Vieb](https://github.com
 The [releases page](https://github.com/Jelmerro/Vieb/releases) also contains the most important changes per release,
 but the list below contains much more technical details.
 
-## Unreleased
+## [Unreleased](https://github.com/Jelmerro/Vieb/compare/0.7.0...master)
 
 ### Added
 
@@ -19,6 +19,8 @@ but the list below contains much more technical details.
 - New tab page with top visited sites listed
 - Setting to disable the added new tab page
 - Setting to automatically enter nav mode in new tabs (or not)
+- Container tabs, which don't share any data with other tabs (such as cookies)
+- Tabs titles are now yellow when they are container tabs
 - Tabs are now bright red when they have crashed
 
 ### Changed
@@ -26,6 +28,10 @@ but the list below contains much more technical details.
 - Moved the setting for new tab position to the newtab collection (newtab.nextToCurrentOne)
 - Reduce duplicate code to execute commands by grouping them in an object
 - Commands can now be entered partially to execute them (if only 1 command matches)
+
+### Removed
+
+- Downloads method settings, because it didn't work out combined with container tabs
 
 ### Fixed